Rename mod to minepi
