+++ minecraft/src/main/java/com/imperialoctopus/minepi/MinePi.java
@@ -1,4 +1,4 @@
-package com.imperialoctopus.freshstart;
+package com.imperialoctopus.minepi;
 
 import net.minecraft.block.Block;
 import net.minecraft.block.Blocks;
@@ -18,13 +18,13 @@ import org.apache.logging.log4j.Logger;
 
 import java.util.stream.Collectors;
 
-@Mod(FreshStart.MODID)
-public class FreshStart {
+@Mod(MinePi.MODID)
+public class MinePi {
 
     private static final Logger LOGGER = LogManager.getLogger();
-    public static final String MODID = "freshstart";
+    public static final String MODID = "minepi";
 
-    public FreshStart() {
+    public MinePi() {
 
         // Register the setup method for modloading
         FMLJavaModLoadingContext.get().getModEventBus().addListener(this::setup);
@@ -44,8 +44,8 @@ public class FreshStart {
 
     private void setup(final FMLCommonSetupEvent event) {
         // some preinit code
-        LOGGER.info("HELLO FROM PREINIT");
-        LOGGER.info("DIRT BLOCK >> {}", Blocks.DIRT.getRegistryName());
+        // LOGGER.info("HELLO FROM PREINIT");
+        // LOGGER.info("DIRT BLOCK >> {}", Blocks.DIRT.getRegistryName());
     }
 
     private void doClientStuff(final FMLClientSetupEvent event) {
@@ -56,23 +56,24 @@ public class FreshStart {
 
     private void enqueueIMC(final InterModEnqueueEvent event) {
         // some example code to dispatch IMC to another mod
-        InterModComms.sendTo(FreshStart.MODID, "helloworld", () -> {
-            LOGGER.info("Hello world from the MDK");
-            return "Hello world";
-        });
+        // InterModComms.sendTo(MinePi.MODID, "helloworld", () -> {
+        // LOGGER.info("Hello world from the MDK");
+        // return "Hello world";
+        // });
     }
 
     private void processIMC(final InterModProcessEvent event) {
         // some example code to receive and process InterModComms from other mods
-        LOGGER.info("Got IMC {}",
-                event.getIMCStream().map(m -> m.getMessageSupplier().get()).collect(Collectors.toList()));
+        // LOGGER.info("Got IMC {}",
+        // event.getIMCStream().map(m ->
+        // m.getMessageSupplier().get()).collect(Collectors.toList()));
     }
 
     // You can use SubscribeEvent and let the Event Bus discover methods to call
     @SubscribeEvent
     public void onServerStarting(FMLServerStartingEvent event) {
         // do something when the server starts
-        LOGGER.info("HELLO from server starting");
+        // LOGGER.info("HELLO from server starting");
     }
 
     // You can use EventBusSubscriber to automatically subscribe events on the
@@ -83,7 +84,7 @@ public class FreshStart {
         @SubscribeEvent
         public static void onBlocksRegistry(final RegistryEvent.Register<Block> blockRegistryEvent) {
             // register a new block here
-            LOGGER.info("HELLO from Register Block");
+            // LOGGER.info("HELLO from Register Block");
         }
     }
 }
