+++ minecraft/src/main/resources/assets/freshstart/blockstates/pi_block_green.json
@@ -1,10 +1,10 @@
 {
   "variants": {
     "lit=false": {
-      "model": "freshstart:block/pi_block_green"
+      "model": "minepi:block/pi_block_green"
     },
     "lit=true": {
-      "model": "freshstart:block/pi_block_green_powered"
+      "model": "minepi:block/pi_block_green_powered"
     }
   }
 }
