+++ minecraft/src/main/resources/assets/freshstart/models/block/pi_block_red_powered.json
@@ -1,11 +1,11 @@
 {
   "parent": "block/cube",
   "textures": {
-    "up": "freshstart:block/pi_top_powered",
-    "down": "freshstart:block/pi_bottom",
-    "east": "freshstart:block/pi_side_red",
-    "west": "freshstart:block/pi_side_red",
-    "north": "freshstart:block/pi_side_red",
-    "south": "freshstart:block/pi_side_red"
+    "up": "minepi:block/pi_top_powered",
+    "down": "minepi:block/pi_bottom",
+    "east": "minepi:block/pi_side_red",
+    "west": "minepi:block/pi_side_red",
+    "north": "minepi:block/pi_side_red",
+    "south": "minepi:block/pi_side_red"
   }
 }
