+++ minecraft/src/main/resources/assets/freshstart/models/block/pi_block_yellow_powered.json
@@ -1,11 +1,11 @@
 {
   "parent": "block/cube",
   "textures": {
-    "up": "freshstart:block/pi_top_powered",
-    "down": "freshstart:block/pi_bottom",
-    "east": "freshstart:block/pi_side_yellow",
-    "west": "freshstart:block/pi_side_yellow",
-    "north": "freshstart:block/pi_side_yellow",
-    "south": "freshstart:block/pi_side_yellow"
+    "up": "minepi:block/pi_top_powered",
+    "down": "minepi:block/pi_bottom",
+    "east": "minepi:block/pi_side_yellow",
+    "west": "minepi:block/pi_side_yellow",
+    "north": "minepi:block/pi_side_yellow",
+    "south": "minepi:block/pi_side_yellow"
   }
 }
