+++ minecraft/src/main/resources/assets/freshstart/models/item/pi_block_red.json
@@ -1,3 +1,3 @@
 {
-  "parent": "freshstart:block/pi_block_red"
+  "parent": "minepi:block/pi_block_red"
 }
