+++ minecraft/src/main/resources/assets/freshstart/models/item/pi_block_yellow.json
@@ -1,3 +1,3 @@
 {
-  "parent": "freshstart:block/pi_block_yellow"
+  "parent": "minepi:block/pi_block_yellow"
 }
